chore: 整理代码结构并修复多处小问题
本次提交包含多项优化和修复：
1. 为gfx库添加futures依赖
2. 拆分midi-loader测试代码到独立文件
3. 重构Root视觉状态，提取到VisualState结构体
4. 修复多处变量引用路径错误
5. 简化代码格式，统一构造函数写法
6. 重构部分导入顺序和代码缩进
7. 添加new_blocking图形上下文创建方法
8. 整理导出库的重新导出结构

+++ crates/export/src/converter.rs
@@ -355,7 +355,7 @@ fn process_dms_track(index: usize, track: &[midly::TrackEvent]) -> crate::dms::D
                 channel.get_or_insert(ch_value);
                 match message {
                     MidiMessage::NoteOn { key, vel } => {
-                        let key_value = u8::from(*key);
+                        let key_value = *key;
                         let vel_value = u8::from(*vel);
                         if vel_value == 0 {
                             if let Some((start_tick, start_vel)) =
@@ -374,7 +374,7 @@ fn process_dms_track(index: usize, track: &[midly::TrackEvent]) -> crate::dms::D
                         }
                     }
                     MidiMessage::NoteOff { key, .. } => {
-                        let key_value = u8::from(*key);
+                        let key_value = *key;
                         if let Some((start_tick, start_vel)) =
                             active_notes.remove(&(ch_value, key_value))
                         {
